Added back update_config_file method.
Added back update_config_file method to fix the loading of an external
geometry file. NWChem looks for the geometry file in the same directory
that the nwchem command was executed. Since we keep the input file in
a separate directory we need to add that directory to the geometry file
path in the input file if it isn't already there.