Creating link to geometry file in job directory.
Rather than editing the load line in the input file to append
the input directory to it we now create a symbolic link to the
geometry file in the job directory to the geometry file in the
input directory. Removed the code that appends the input directory
to the load line in the input file. Tested with the runner script
and the UI and everything appears to be working.