Using docker container for nwchem json conversion.
Switched to running the json conversion script for nwchem in a Docker
container. The nwchem tasks now copies the output file to the server,
runs the docker container to generate the json, and then copies the
json output back to the cluster. Several things need to be cleaned
up like the hard-coded settings for the docker container.