Use specified number of CPUs in the visualizer
For slurm, the number of CPUs needs to be specified in order to use more
than one. Additionally, the [cumulus slurm template](https://github.com/Kitware/cumulus/blob/27a66935e4ae1bb77b5ee3398928739f147785e3/cumulus/templates/schedulers/slurm.sh#L11-L13) requires both the
number of nodes and the number of CPUs to be specified before it
will provide to slurm the --cpus-per-task flag.

Thus, we should ensure both the number of nodes and the number
of CPUs are specified.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>